refactor(core): 🔨 split DomainHealthCheck into partial files
